"target" or "gmcp.Char.Status.target" as fallback
If the autotargeter found no valid targets, the target variable or, if
target is not filled, the gmcp.Char.Status.target value.

There is currently an issue with the gmcp fallback though: if set,
arguments to the kill command are not used, instead only the target is
evaluated. I am not sure if this is a bug or intended feature.

More future improvements contain, but are not limited to, automatic
updating of the target list, if new targets are added to the list of
possible targets.