Change current target list wit additional prio
The current target list is now updated whenever a new item is added to
the priority list. This is not echoed to the client yet though!

Also fixed a bug that made it not echo the updated target list, if new
and old list were different but same length.

Save prio list only when changing instead of on each kill.

Formatting changes.