Detach from svo step 1: create strategy structure
We use a pattern that is very similar to the strategy pattern in OOP, which makes it very easy to switch the underlying system, even on the fly.

For that I created a structure that will hold system dependant code. For now all calls are hardwired to the svo "strategy".