Add battlerage to the basher
This commit introduces first support for battlerage. It allows for affliction
tracking of denizens (no triggers though). It also adds a simple strategy to use
with battlerage.

A known problem is shielding of denizens with wundersys. The version will NOT be
published without resolving this as it could possibly be the cause of death.